requirements.txt
asgiref==3.6.0
Django==4.2.1
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow==9.5.0
sqlparse==0.4.4
